Move these useful functions into places that make a bit more sense
